+++ src/com/event.rs
@@ -1,18 +1,14 @@
 use futures::task::{self, ArcWake};
 use futures::Stream;
+use windows::Win32::Foundation::{CloseHandle, HANDLE};
+use windows::Win32::System::Com::{CoWaitForMultipleObjects, CWMO_DISPATCH_CALLS};
+use windows::Win32::System::Threading::{CreateEventW, SetEvent};
+use windows::Win32::System::WindowsProgramming::INFINITE;
 
 use std::pin::Pin;
+use std::ptr;
 use std::sync::Arc;
 use std::task::{Context, Poll};
-use std::{mem::MaybeUninit, ptr};
-
-use winapi::um::combaseapi::{CoWaitForMultipleObjects, CWMO_DISPATCH_CALLS};
-use winapi::um::handleapi::CloseHandle;
-use winapi::um::synchapi::{CreateEventW, SetEvent};
-use winapi::um::winbase::INFINITE;
-use winapi::um::winnt::HANDLE;
-
-use crate::hresult::check;
 
 struct ComWaker {
     ready_event: HANDLE,
@@ -28,7 +24,7 @@ impl ArcWake for ComWaker {
 
 impl Drop for ComWaker {
     fn drop(&mut self) {
-        unsafe { CloseHandle(self.ready_event as *mut _) };
+        unsafe { CloseHandle(self.ready_event) };
     }
 }
 
@@ -37,22 +33,14 @@ unsafe impl Sync for ComWaker {}
 
 impl ComWaker {
     pub fn new() -> Self {
-        let ready_event = unsafe { CreateEventW(ptr::null_mut(), 0, 0, ptr::null_mut()) };
+        let ready_event = unsafe { CreateEventW(ptr::null_mut(), false, false, None).unwrap() };
         Self { ready_event }
     }
 
     pub fn sleep(&self) {
         unsafe {
-            let mut which = MaybeUninit::uninit();
-            check(CoWaitForMultipleObjects(
-                CWMO_DISPATCH_CALLS,
-                INFINITE,
-                1,
-                &self.ready_event,
-                which.as_mut_ptr(),
-            ))
-            .expect("failed to wait for wake");
-            which.assume_init();
+            CoWaitForMultipleObjects(CWMO_DISPATCH_CALLS.0 as u32, INFINITE, &[self.ready_event])
+                .expect("failed to wait for wake");
         }
     }
 }