+++ src/soundcore/error.rs
@@ -1,14 +1,13 @@
 use std::error::Error;
 use std::fmt;
 
-use crate::hresult::Win32Error;
 use crate::media::GetPropertyError;
 
 /// Describes an error that occurred while acting on Creative's SoundCore API.
 #[derive(Debug)]
 pub enum SoundCoreError {
     /// Some Win32 error occurred.
-    Win32(Win32Error),
+    Win32(windows::core::Error),
     /// The specified device does not support implement the SoundCore API.
     NotSupported,
 }
@@ -31,8 +30,8 @@ impl Error for SoundCoreError {
     }
 }
 
-impl From<Win32Error> for SoundCoreError {
-    fn from(err: Win32Error) -> SoundCoreError {
+impl From<windows::core::Error> for SoundCoreError {
+    fn from(err: windows::core::Error) -> SoundCoreError {
         SoundCoreError::Win32(err)
     }
 }