+++ src/soundcore/parameter.rs
@@ -1,12 +1,12 @@
-use std::mem::{self, MaybeUninit};
+use std::mem;
+use std::mem::MaybeUninit;
 use std::str;
 
 use tracing::{info, trace, trace_span};
-use winapi::shared::winerror::E_ACCESSDENIED;
+use windows::Win32::Foundation::E_ACCESSDENIED;
 
 use crate::com::ComObject;
 use crate::ctsndcr::{ISoundCore, Param, ParamInfo, ParamValue};
-use crate::hresult::{check, Win32Error};
 
 /// Captures the value of a parameter.
 #[derive(Clone, Copy, Debug)]
@@ -83,7 +83,7 @@ impl SoundCoreParameter {
     ///
     /// May return `Err(Win32Error { code: E_ACCESSDENIED })` when getting a
     /// parameter that is not currently applicable.
-    pub fn get(&self) -> Result<SoundCoreParamValue, Win32Error> {
+    pub fn get(&self) -> windows::core::Result<SoundCoreParamValue> {
         // varsize -> not supported
         if self.kind == 5 {
             return Ok(SoundCoreParamValue::None);
@@ -94,14 +94,12 @@ impl SoundCoreParameter {
                 feature: self.feature_id,
                 param: self.id,
             };
-            let mut value = MaybeUninit::uninit();
-            let span = trace_span!(
-                "Fetching parameter value .{context}.{feature_id}.{id}...",
-            );
+            let span = trace_span!("Fetching parameter value .{context}.{feature_id}.{id}...",);
             let _span = span.enter();
-            match check(self.core.GetParamValue(param, value.as_mut_ptr())) {
-                Ok(_) => {}
-                Err(Win32Error { code, .. }) if code == E_ACCESSDENIED => {
+            let mut value = MaybeUninit::uninit();
+            let value = match self.core.GetParamValue(param, value.as_mut_ptr()).ok() {
+                Ok(()) => value.assume_init(),
+                Err(error) if error.code() == E_ACCESSDENIED => {
                     trace!(
                         "Got parameter value .{}.{}.{} = {}",
                         self.context,
@@ -113,7 +111,6 @@ impl SoundCoreParameter {
                 }
                 Err(error) => return Err(error),
             };
-            let value = value.assume_init();
             span.record("value", &tracing::field::debug(&value));
             Ok(convert_param_value(&value))
         }
@@ -122,7 +119,7 @@ impl SoundCoreParameter {
     ///
     /// May return `Err(Win32Error { code: E_ACCESSDENIED })` when setting a
     /// parameter that is not currently applicable.
-    pub fn set(&mut self, value: &SoundCoreParamValue) -> Result<(), Win32Error> {
+    pub fn set(&mut self, value: &SoundCoreParamValue) -> windows::core::Result<()> {
         unsafe {
             let param = Param {
                 context: self.context,
@@ -155,8 +152,7 @@ impl SoundCoreParameter {
                 "Setting {}.{} = {:?}",
                 self.feature_description, self.description, value
             );
-            check(self.core.SetParamValue(param, param_value))?;
-            Ok(())
+            self.core.SetParamValue(param, param_value).ok()
         }
     }
 }